+++ lib/grinder.dart
@@ -1,13 +1,11 @@
 // Copyright 2015 Google. All rights reserved. Use of this source code is
 // governed by a BSD-style license that can be found in the LICENSE file.
 
-/**
- * Dart workflows, automated.
- *
- * See the [README][] for detailed usage information.
- *
- * [README]: https://pub.dartlang.org/packages/grinder
- */
+/// Dart workflows, automated.
+///
+/// See the [README][] for detailed usage information.
+///
+/// [README]: https://pub.dartlang.org/packages/grinder
 library grinder;
 
 import 'dart:async';
@@ -45,7 +43,7 @@ export 'src/task_invocation.dart';
 Future grind(List<String> args, {bool verifyProjectRoot: true}) {
   try {
     discoverTasks(grinder, currentMirrorSystem().isolate.rootLibrary);
-    return handleArgs(args, verifyProjectRoot: verifyProjectRoot);
+    return runTasks(args, verifyProjectRoot: verifyProjectRoot);
   } catch (e) {
     if (e is GrinderException) {
       fail(e.message);
@@ -55,29 +53,18 @@ Future grind(List<String> args, {bool verifyProjectRoot: true}) {
   }
 }
 
-/**
- * Start the build process. This should be called at the end of the `main()`
- * method. If there is a task failure, this method will halt task execution and
- * throw.
- */
+/// Start the build process. This should be called at the end of the `main()`
+/// method. If there is a task failure, this method will halt task execution and
+/// throw.
 @Deprecated('Use `grind` instead.')
 Future startGrinder(List<String> args, {bool verifyProjectRoot: true}) {
-  return handleArgs(args, verifyProjectRoot: verifyProjectRoot);
+  return runTasks(args, verifyProjectRoot: verifyProjectRoot);
 }
 
-/// Used to define a method body for a task. Note: a task's context is now
-/// available as a global variable ('context'). Your task functions should no
-/// longer be definied with a single `GrinderContext` parameter.
-@Deprecated('''Use a nullary function instead.  A task's context can now be
-accessed via the top-level `context` getter.''')
-typedef dynamic TaskFunction(GrinderContext context);
-
-/**
- * Add a new task definition to the global [Grinder] instance. A [name] is
- * required. If specified, a [taskFunction] is invoked when the task starts.
- * Any dependencies of the task, that need to run before it, should be passed
- * in via [depends].
- */
+/// Add a new task definition to the global [Grinder] instance. A [name] is
+/// required. If specified, a [taskFunction] is invoked when the task starts.
+/// Any dependencies of the task, that need to run before it, should be passed
+/// in via [depends].
 @Deprecated('Use the task annotations instead.')
 void task(String name,
     [Function taskFunction, List<String> depends = const []]) {