+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # grinder.dart changes
 
+## 0.8.1
+- Added the ability to pass parameters to tasks from the command line (using
+  a `TaskArgs` parameter on task functions)
+- Removed the older way to pass flags and options to tasks (via named and
+  positional parameters on task functions)
+- Remove a dependency on a non-strong mode package (`unscripted`)
+
 ## 0.8.0+3
 - All APIs, including `pub run grinder`, always use the same Dart version as the
   user.